requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.54.4
pdfplumber==0.11.4
pymupdf==1.24.11
python-docx==1.1.2
python-multipart==0.0.12
python-dotenv==1.0.1
